[4.4] Move the Cypress Tests to ESM (#43676)
* [4.4] Move the Cypress Tests to ESM

* drone  signature

* resolve conflicts with #43663 - part 2

* Fix path.join

* Fix path.join - 2

* Fix import join from path

* Remove the forgotten commented out orginial imports from plugin/index.js

---------
